StudentDetailController moved to new folder Student 	renamed:    app/Http/Controllers/api/StudentDetailController.php -> app/Http/Controllers/api/Student/StudentDetailController.php 	modified:   routes/api/v1.php 	modified:   tests/Feature/Student/StudentDetailControllerTest.php
