+++ tests/Feature/Student/StudentDetailControllerTest.php
@@ -6,7 +6,7 @@ namespace Tests\Feature\Student;
 
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Tests\TestCase;
-use App\Http\Controllers\api\StudentDetailController;
+use App\Http\Controllers\api\Student\StudentDetailController;
 use App\Service\StudentDetailService;
 use App\Models\Student;
